refactor: optimize bunner-init and isScriptDirectoryInitialized for improved readability and performance

+++ bunner-commands/bunner-init.ts
@@ -1,4 +1,5 @@
-import { writeFile, chmod } from 'node:fs/promises';
+import { existsSync } from 'node:fs';
+import { chmod, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 
 import { $ } from 'bun';
@@ -29,22 +30,25 @@ export default defineCommand({
     action: async ({ options }) => {
         const templateFiles: string[] = [];
 
-        (await eachTemplateFile()).forEach(
-            async ({ templateFileBaseName, templateFileTransformed }) => {
+        for (const { templateFileBaseName, templateFileTransformed } of await eachTemplateFile()) {
+            const targetPath = join(options.directory, templateFileBaseName);
+            templateFiles.push(templateFileBaseName);
+
+            if (!existsSync(targetPath)) {
                 log.info(`Creating template file: ${templateFileBaseName} in ${options.directory}`);
-                templateFiles.push(templateFileBaseName);
-                const targetPath = join(options.directory, templateFileBaseName);
                 await writeFile(targetPath, templateFileTransformed);
                 await chmod(targetPath, 0o666); // Set 666 permissions
-            },
-        );
+            }
+        }
 
         // Generate .gitignore file
-        log.info(`Creating .gitignore file in ${options.directory}`);
-        const gitignoreContent = templateFiles.join('\n');
         const gitignorePath = join(options.directory, '.gitignore');
-        await writeFile(gitignorePath, gitignoreContent);
-        await chmod(gitignorePath, 0o666); // Set 666 permissions on .gitignore
+        if (!existsSync(gitignorePath)) {
+            log.info(`Creating .gitignore file in ${options.directory}`);
+            const gitignoreContent = templateFiles.join('\n');
+            await writeFile(gitignorePath, gitignoreContent);
+            await chmod(gitignorePath, 0o666); // Set 666 permissions on .gitignore
+        }
 
         if (options.install) {
             log.info(
